Hi. I have been looking at this code for more than 3 years, and finally I find a way to improve the performance. This consists of two commits, of which you may choose to merge only one.

### Base performance
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      2.080 s ±  0.018 s    [User: 1.956 s, System: 0.123 s]
  Range (min … max):    2.055 s …  2.105 s    10 runs
```

### Improve code layout

Commit one groups similar functionality in the parse code. This makes the code easier to read. As well as limits the amount of '+1', '-1', etc. actions. We don't create new name, sequence and qualities variables until it is certain that the FASTQ record has the complete 4 newlines.
This code is slightly faster (5-10% in my benchmarks). Maybe this has also to do with the branch predictor in the CPU having an easier time because similar functionality is grouped. (EDIT: Or this may be the layout issue mentioned below in which case it is a very small increment probably)
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.878 s ±  0.014 s    [User: 1.769 s, System: 0.109 s]
  Range (min … max):    1.855 s …  1.907 s    10 runs
```

### Infer quality length from sequence length

I was reading [joelonsoftware.com's article on microsoft office formats](https://www.joelonsoftware.com/2008/02/19/why-are-the-microsoft-office-file-formats-so-complicated-and-some-workarounds/), and he noted that binary formats are much easier to read, because you can simply copy whole parts in memory directly. That is also what happens in this parser, which is good. I was lamenting that it is unfortunate that we do not know the length beforehand because FASTQ does not encode that.

Then I realized the spec says that the qualities string MUST be the length of the sequence string. So it **does** encode one length.

For each character in the input, two checks were performed. Is it a newline, and have we reached bufend yet?
With this change the length of the quality string is inferred from the sequence string. Therefore we can skip
sequence length x 2 checks. We just perform checks afterward to see if we end on a newline.

This has a slight risk, as there are some possible outcomes:
1. The newline is indeed the last newline of the record.
2. The newline is a newline in a following record, because the quality string was truncated and we were unlucky.

In case of 1. Success, but in case of 2. there are two possible outcomes:
a. It is the ending newline of the name, sequence or second header line.
   In which case the parser will throw an error for the next record.
   Which is good, problem averted.
b. It is the ending newline of the next record's quality string. In that case
   we have included an entire record in this record's quality string and the
   parser keeps happily parsing.

In case of b, this leads to a missing record, and a record with
invalid quality values. This might be catched by upstream tools
(problem averted!) or even by cutadapt itself (in paired mode), or it will lead to wrong quality values for this record.
Probably one misrepresented record will not lead to a different analysis outcome.
 If there are multiple records with an erroneous quality length, they are bound to be not lucky all the time and trigger parser erros.

So with this commit we sacrifice fidelty in an extremely rare edge case error for speed.  Dnaio does this all the time though: quality and sequence are not checked for if the characters are in the valid ASCII range. So this is in the spirit of the project. EDIT: to better phrase this. Garbage in == garbage out. With this code change the ability is lost to catch one very rare case of garbage. With correct input the output will still always be correct.

And now the speed improvement:

```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.310 s ±  0.017 s    [User: 1.182 s, System: 0.128 s]
  Range (min … max):    1.278 s …  1.338 s    10 runs
``` 

### Disclaimer

Compiling C is prone to layout problems which can make the performance differ as much as 40%. See this very interesting video below. I did not benchmark this change with the rigor as shown in this video. However because the code changes do demonstrably less work each, I think the performance gains are real.
https://www.youtube.com/watch?v=7g1Acy5eGbE

EDIT: After deleting two unused variables for instance after recompilation the speed improvement is:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.462 s ±  0.019 s    [User: 1.331 s, System: 0.130 s]
  Range (min … max):    1.438 s …  1.502 s    10 runs
```
Still good, but a 150 ms difference!  For compiler and layout things that are not easy to fix.