@rhpvorderman 2021-12-09
@marcelm You will **love** this. Instead of using C code to find the newlines I used `memchr`. A lot of the `mem*` code in glibc is implemented in assembler.

And there we go:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.069 s ±  0.013 s    [User: 946.3 ms, System: 122.3 ms]
  Range (min … max):    1.047 s …  1.093 s    10 runs
```