@marcelm 2021-12-09
Oh, really cool! I would love to look at this right now, but that will probably have to wait until next week, just so you know.