@rhpvorderman 2021-12-09
No problem! I also checked out the memchr improvement while not using the quality length hack:
```
# No quality skip but with memchr
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.139 s ±  0.014 s    [User: 1.028 s, System: 0.111 s]
  Range (min … max):    1.125 s …  1.173 s    10 runs
```
So the performance savings are not as big when using memchr. It might be considered not doing it. I will make an alternative PR for this.