@rhpvorderman 2021-12-12
I was thinking of a couple of ways to improve the parser still, but I thought it would be instructive to have a baseline benchmark for performance.

So I wrote this script:
```python3
#!/usr/bin/env python3

import sys

with open(sys.argv[1], "rb")  as file_h:
    for line in file_h:
        pass
```
This does almost exactly the same thing as dnaio. It reads in chunks of files and finds the line ends using memchr. It then copies the found lines directly into python objects. (It skips the encoding part, but we have already established that encoding with latin-1 is the same as copying it directly into a python object).
This creates exactly the same amount of python objects per FASTQ record (4 bytes objects vs 3 unicode objects and 1 custom class). The mechanisms are also mostly the same as CPython uses memchr internally.

The results:
```
Benchmark #1: python dnaio_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.104 s ±  0.025 s    [User: 997.1 ms, System: 106.9 ms]
  Range (min … max):    1.069 s …  1.158 s    10 runs

Benchmark #1: python binary_read.py ~/test/big2.fastq
  Time (mean ± σ):      1.004 s ±  0.038 s    [User: 872.2 ms, System: 132.1 ms]
  Range (min … max):    0.974 s …  1.085 s    10 runs
```

So the iteration of dnaio as presented in this PR is only 10% slower than the baseline performance. It will be impossible to get it below 1 second using python's file reading mechanisms. Optimizations to the parser code are moot after this point.
Significant gains can only be made by improving the file reading. Using mmaps is a possibility, but since we use gzipped fastq files almost exclusively in production, that will not work. And unpacking first and then parsing is slower anyway. (And a huge disk IO burner).