@marcelm 2021-12-13
I’m reviewing this commit by commit, please let me know beforehand if you want to do any force-pushing.

First commit looks fine, nice improvement.