@rhpvorderman 2021-12-14
If it helps the review process. #15 is basically the improvement of the parser using `memchr`. The only difference with this branch is this commit:
https://github.com/rhpvorderman/dnaio/commit/3c0edf8dffee65d49b4444d083c26e998c93d5a6

This commit highlights how I implemented the quality length shortcut.