@marcelm 2021-12-16
> EDIT: Although, of course I am in favour for not leaving the performance on the table. But I do not want to force the quality length shortcut and sugarcoat it with performance that is not derived from it.

I’ve thought about this a bit, and my tendency at the moment is to not include this shortcut. I think it’s smart and of course it makes things a bit faster, but it’s still a hack. We’re still working with Python, so there’s already a lot of performance left on the table, as you say, so I would tend to try to optimize other parts of the pipeline (of which dnaio is likely only a small part), instead.