@rhpvorderman 2021-12-16
Yes, it seems to save only 80ms on 5 million illumina records, so that is not very stellar.

>  (of which dnaio is likely only a small part),

I am currently working on a fastq-filter program, written in Python + Cython, and dnaio is the vast majority of the compute time (80-90%). Hence my improvements here.

Could you have a look at #15 instead? That improves the speed by almost 100% by using memchr. It also reorganizes the fastq_iter code a bit to be more readable. It is code without shorcuts.
Thank you for your feedback so far.