@marcelm 2021-12-16
> Could you have a look at #15 instead?

Yes of course! The rest is fine and unless I find something strange, I want to merge it.